Fixed SysEx send and Msg receive logic

+++ windows/MidiHeaderFlags.cs
@@ -1,14 +1,14 @@
 ﻿namespace LpxApi.windows;
 
-public readonly struct MidiHeaderFlags(ulong value)
+public readonly struct MidiHeaderFlags(uint value)
 {
-    public ulong Value { get; } = value;
+    public uint Value { get; } = value;
 
     public bool Done => (Value & 0x1) == 0x1;
     public bool Prepared => (Value & 0x2) == 0x2;
     public bool InQueue => (Value & 0x4) == 0x4;
     public bool IsStream => (Value & 0x8) == 0x8;
 
-    public static implicit operator MidiHeaderFlags(ulong l) => new(l);
-    public static implicit operator ulong(MidiHeaderFlags f) => f.Value;
+    public static implicit operator MidiHeaderFlags(uint l) => new(l);
+    public static implicit operator uint(MidiHeaderFlags f) => f.Value;
 }
